* Fixed: Possible runtime assert
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@9017 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/sdk/projectfileoptionsdlg.cpp
@@ -279,7 +279,7 @@ void ProjectFileOptionsDlg::EndModal(int retCode)
     if (retCode == wxID_OK && m_ProjectFile)
     {
         wxCheckListBox *list = XRCCTRL(*this, "lstTargets", wxCheckListBox);
-        for (int i = 0; i < (int)list->GetCount(); i++)
+        for (unsigned int i = 0; i < list->GetCount(); i++)
         {
             if (list->IsChecked(i))
                 m_ProjectFile->AddBuildTarget(list->GetString(i));
